pyproject.toml
[project]
name = "proj"
version = "0.1.0"
dependencies = [
    "flask>=3.1.1",
    "line-bot-sdk>=3.17.1",
    "python-dotenv",
]
